02_Website-Builder/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
